+++ smok/logging.py
@@ -73,8 +73,11 @@ class SMOKLogHandler(Handler, Monitor):
         if record.exc_info:
             f = frame_from_traceback(record.exc_info[2])
             tb = Traceback(f)
-            tb_json = base64.b64encode(
-                gzip.compress(ujson.dumps(tb.to_json()).encode('utf-8'), 9)).decode('utf-8')
+            try:
+                tb_json = base64.b64encode(
+                    gzip.compress(ujson.dumps(tb.to_json()).encode('utf-8'), 9)).decode('utf-8')
+            except OverflowError:
+                return None
             dct.update(exception_text=tb.pretty_format(),
                        exception_traceback=tb_json)
 
@@ -82,7 +85,9 @@ class SMOKLogHandler(Handler, Monitor):
 
     def emit(self, record: LogRecord) -> None:
         if self.device.log_publisher.queue.qsize() < MAX_LOG_BUFFER_SIZE:
-            self.device.log_publisher.queue.put(self.record_to_json(record))
+            le = self.record_to_json(record)
+            if le is not None:
+                self.device.log_publisher.queue.put(le)
 
     def close(self) -> None:
         pass