+++ smok/sync_workers/ngtt.py
@@ -5,28 +5,24 @@ import typing as tp
 from ngtt.exceptions import ConnectionFailed, DataStreamSyncFailed
 from ngtt.orders import Order
 from smok.pathpoint.orders import sections_from_list
-from smok.sync_workers.base import BaseSyncWorker, SyncError
+from smok.sync_workers.base import SyncError
 from ngtt.uplink import NGTTConnection
+from smok.sync_workers.http import HTTPSyncWorker
 
 logger = logging.getLogger(__name__)
 
 
-class NGTTSyncWorker(BaseSyncWorker):
+class NGTTSyncWorker(HTTPSyncWorker):
     """
     Note that this is ready to start when .executor on the device is ready
     """
 
-    def sync_logs(self, data: tp.List[dict]):
-        try:
-            self.connection.stream_logs(data)
-        except ConnectionFailed as e:
-            raise SyncError(e.is_due_to_no_internet) from e
-
     def __init__(self, device: 'SMOKDevice'):
-        super().__init__(device, True)
+        super().__init__(device)
         self.connection = NGTTConnection(device.temp_file_for_cert,
                                          device.temp_file_for_key,
                                          None if self.device.dont_do_orders else self.process_orders)
+        self.has_async_orders = True
         time.sleep(5)  # Give it some time to connect
 
     def process_orders(self, orders: Order):