+++ smok/extras/pp_database/base.py
@@ -77,6 +77,7 @@ class BasePathpointDatabase(metaclass=ABCMeta):
     def checkpoint(self) -> None:
         """
         Called by the communicator thread, once every about 20 seconds.
+        May be called much more often, it's the function responsibility to throttle.
         """
 
     @abstractmethod