+++ docs/changelog.rst
@@ -5,6 +5,7 @@ v0.16.3
 ~~~~~~~
 
 * improved logging for NGTT
+* improved behaviour of the logging subsystem on low memory conditions
 
 v0.16.2
 ~~~~~~~