+++ docs/changelog.rst
@@ -9,6 +9,7 @@ v0.16.3
 * improved behaviour of connection resetting
 * fixed a bug that wouldn't allow smok to run on Python 3.5
 * HTTP API will use minijson everywhere
+* logs won't be synced via NGTT anymore - they are dropped too often
 
 v0.16.2
 ~~~~~~~